requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
gunicorn==23.0.0
iniconfig==2.1.0
kombu==5.5.3
mccabe==0.7.0
packaging==25.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
